src/services/events/requirements.txt
websockets
aiohttp
python-dotenv
pydantic
redis
aio-pika